Moved Payment History Charge and Helper into payments packages
