+++ app/views/payments/openPayments.scala.html
@@ -49,7 +49,7 @@
               <thead>
                   <tr>
                       <th scope="col">
-                          <div class="visuallyhidden">@messages("openPayments.action")</div>
+                          <div class="visuallyhidden">@messages("openPayments.description")</div>
                       </th>
 
                       <th scope="col" class="numeric">
@@ -57,7 +57,7 @@
                       </th>
 
                       <th scope="col">
-                          <div class="visuallyhidden">@messages("openPayments.description")</div>
+                          <div class="visuallyhidden">@messages("openPayments.action")</div>
                       </th>
                   </tr>
               </thead>
