+++ app/views/payments/paymentHistory.scala.html
@@ -62,7 +62,7 @@
             </thead>
             <tbody>
             @for(transaction <- model.transactions) {
-                @views.html.templates.paymentsHistoryCharge(transaction)
+                @views.html.templates.payments.paymentsHistoryCharge(transaction)
             }
             </tbody>
